requirements.txt
streamlit
numpy
opencv-python
Pillow
scikit-learn
matplotlib
